requirements-dev.txt
green
flake8
ipython
vcrpy
